+++ src/lib/FakeDOMStringList.ts
@@ -1,73 +1,79 @@
-class FakeDOMStringList extends Array<string> {
+const regex = /^\d+$/;
+const isPositiveInteger = (prop: unknown): prop is string => {
+    return typeof prop === "string" && regex.test(prop);
+};
+
+class FakeDOMStringList implements DOMStringList {
+    private _values: string[];
+
+    constructor(...values: string[]) {
+        this._values = values;
+
+        return new Proxy(this, {
+            get(target, prop) {
+                if (isPositiveInteger(prop)) {
+                    const index = parseInt(prop);
+                    return target._values[index];
+                }
+                return (target as any)[prop];
+            },
+            has(target, prop) {
+                if (isPositiveInteger(prop)) {
+                    const index = parseInt(prop);
+                    return index < target._values.length;
+                }
+                return prop in target;
+            },
+            ownKeys(target) {
+                return Array.from({ length: target._values.length }, (_, i) =>
+                    String(i),
+                );
+            },
+            getOwnPropertyDescriptor(target, prop) {
+                if (isPositiveInteger(prop)) {
+                    const index = parseInt(prop);
+                    if (index < target._values.length) {
+                        return {
+                            value: target._values[index],
+                            enumerable: true,
+                            configurable: true,
+                            writable: false,
+                        };
+                    }
+                }
+            },
+        });
+    }
+
     contains(value: string) {
-        for (const value2 of this) {
-            if (value === value2) {
-                return true;
-            }
-        }
-        return false;
+        return this._values.includes(value);
     }
 
     item(i: number) {
-        if (i < 0 || i >= this.length) {
+        if (i < 0 || i >= this._values.length) {
             return null;
         }
-        return this[i];
+        return this._values[i];
     }
+
+    get length() {
+        return this._values.length;
+    }
+
+    [Symbol.iterator]() {
+        return this._values[Symbol.iterator]();
+    }
+
+    // Handled by proxy
+    [index: number]: string;
 
     // Used internally, should not be used by others. I could maybe get rid of these and replace rather than mutate, but too lazy to check the spec.
     _push(...values: Parameters<typeof Array.prototype.push>) {
-        return Array.prototype.push.call(this, ...values);
+        return this._values.push(...values);
     }
     _sort(...values: Parameters<typeof Array.prototype.sort>) {
-        return Array.prototype.sort.call(this, ...values);
+        return this._values.sort(...values);
     }
 }
-
-// Would be nice to remove these properties to fix https://github.com/dumbmatter/fakeIndexedDB/issues/66 but for some reason it breaks Dexie - see test/dexie.js and FakeDOMStringList tests
-/*
-// From https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array
-const arrayPropertiesToDelete = ["from", "isArray", "of"];
-const arrayMethodsToDelete = [
-    "at",
-    "concat",
-    "copyWithin",
-    "entries",
-    "every",
-    "fill",
-    "filter",
-    "find",
-    "findIndex",
-    "flat",
-    "flatMap",
-    "forEach",
-    "includes",
-    "indexOf",
-    "join",
-    "keys",
-    "lastIndexOf",
-    "map",
-    "pop",
-    "push",
-    "reduce",
-    "reduceRight",
-    "reverse",
-    "shift",
-    "slice",
-    "some",
-    "sort",
-    "splice",
-    "unshift",
-    "values",
-];
-
-// Set to undefined rather than delete, so it doesn't go up the chain to Array. Not perfect, but good enough?
-for (const property of arrayPropertiesToDelete) {
-    (FakeDOMStringList as any)[property] = undefined;
-}
-for (const property of arrayMethodsToDelete) {
-    (FakeDOMStringList as any).prototype[property] = undefined;
-}
-*/
 
 export default FakeDOMStringList;